@@ -1,32 +1,32 @@
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_image/firebase_image.dart';
 import 'package:firebase_image/src/firebase_image.dart';
 import 'package:firebase_image/src/image_object.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:sqflite/sqflite.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:sqflite/sqflite.dart';
 
 class FirebaseImageCacheManager {
   static const String key = 'firebase_image';
 
-  Database db;
+  late Database db;
   static const String dbName = '$key.db';
   static const String table = 'images';
-  String basePath;
+  late String basePath;
 
   final CacheRefreshStrategy cacheRefreshStrategy;
 
   FirebaseImageCacheManager(
-    this.cacheRefreshStrategy,
-  );
+      this.cacheRefreshStrategy,
+      );
 
   Future<void> open() async {
     db = await openDatabase(
-      join(await getDatabasesPath(), dbName),
+      join((await getDatabasesPath())!, dbName),
       onCreate: (Database db, int version) async {
         await db.execute('''
           CREATE TABLE $table (
@@ -76,7 +76,7 @@ class FirebaseImageCacheManager {
     return maps.length > 0;
   }
 
-  Future<FirebaseImageObject> get(String uri, FirebaseImage image) async {
+  Future<FirebaseImageObject?> get(String uri, FirebaseImage image) async {
     final List<Map<String, dynamic>> maps = await db.query(
       table,
       columns: const [
@@ -90,7 +90,7 @@ class FirebaseImageCacheManager {
     );
     if (maps.length > 0) {
       FirebaseImageObject returnObject =
-          FirebaseImageObject.fromMap(maps.first);
+      FirebaseImageObject.fromMap(maps.first);
       returnObject.reference = getImageRef(returnObject, image.firebaseApp);
       if (CacheRefreshStrategy.BY_METADATA_DATE == this.cacheRefreshStrategy) {
         checkForUpdate(returnObject, image); // Check for update in background
@@ -100,16 +100,16 @@ class FirebaseImageCacheManager {
     return null;
   }
 
-  Reference getImageRef(FirebaseImageObject object, FirebaseApp firebaseApp) {
+  Reference getImageRef(FirebaseImageObject object, FirebaseApp? firebaseApp) {
     FirebaseStorage storage =
-        FirebaseStorage.instanceFor(app: firebaseApp, bucket: object.bucket);
+    FirebaseStorage.instanceFor(app: firebaseApp, bucket: object.bucket);
     return storage.ref().child(object.remotePath);
   }
 
   Future<void> checkForUpdate(
       FirebaseImageObject object, FirebaseImage image) async {
-    int remoteVersion =
-        (await object.reference.getMetadata()).updated.millisecondsSinceEpoch;
+    int remoteVersion = (await object.reference.getMetadata())
+        .updated?.millisecondsSinceEpoch ?? -1;
     if (remoteVersion != object.version) {
       // If true, download new image for next load
       await this.upsertRemoteFileToCache(object, image.maxSizeBytes);
@@ -131,25 +131,25 @@ class FirebaseImageCacheManager {
     );
   }
 
-  Future<Uint8List> localFileBytes(FirebaseImageObject object) async {
+  Future<Uint8List?> localFileBytes(FirebaseImageObject? object) async {
     if (await _fileExists(object)) {
-      return File(object.localPath).readAsBytes();
+      return File(object!.localPath!).readAsBytes();
     }
     return null;
   }
 
-  Future<Uint8List> remoteFileBytes(
+  Future<Uint8List?> remoteFileBytes(
       FirebaseImageObject object, int maxSizeBytes) {
     return object.reference.getData(maxSizeBytes);
   }
 
-  Future<Uint8List> upsertRemoteFileToCache(
+  Future<Uint8List?> upsertRemoteFileToCache(
       FirebaseImageObject object, int maxSizeBytes) async {
     if (CacheRefreshStrategy.BY_METADATA_DATE == this.cacheRefreshStrategy) {
-      object.version =
-          (await object.reference.getMetadata()).updated.millisecondsSinceEpoch;
+      object.version = (await object.reference.getMetadata())
+          .updated?.millisecondsSinceEpoch ?? 0;
     }
-    Uint8List bytes = await remoteFileBytes(object, maxSizeBytes);
+    Uint8List? bytes = await remoteFileBytes(object, maxSizeBytes);
     await putFile(object, bytes);
     return bytes;
   }
@@ -165,11 +165,11 @@ class FirebaseImageCacheManager {
     return await upsert(object);
   }
 
-  Future<bool> _fileExists(FirebaseImageObject object) async {
+  Future<bool> _fileExists(FirebaseImageObject? object) async {
     if (object?.localPath == null) {
       return false;
     }
-    return File(object.localPath).exists();
+    return File(object!.localPath!).exists();
   }
 
   Future<String> _createFilePath() async {
